requirements.txt
docker
cx_Oracle
distlib==0.3.6
filelock==3.9.0
virtualenv==20.17.1
kafka-python==2.0.2
pykafka==2.8.0